ENH: Include MONAILabelReviewerLogic class in main module file
It is Slicer standard to include the Widget, Logic, and Test class for a given module in the same {ModuleName}.py file.
